requirements.txt
beautifulsoup4-4.8.2 
bs4-0.0.1 
soupsieve-2.0
certifi-2019.11.28 
chardet-3.0.4 
idna-2.9 
requests-2.23.0 
urllib3-1.25.8
numpy-1.18.2 
pandas-1.0.3 
python-dateutil-2.8.1 
pytz-2019.3 
six-1.14.